+++ runtime/src/utils.rs
@@ -1,0 +1,36 @@
+use crate::*;
+
+impl<C> frame_system::offchain::SendTransactionTypes<C> for Runtime
+where
+	RuntimeCall: From<C>,
+{
+	type Extrinsic = UncheckedExtrinsic;
+	type OverarchingCall = RuntimeCall;
+}
+
+pub struct OnChainSeqPhragmen;
+impl onchain::Config for OnChainSeqPhragmen {
+	type System = Runtime;
+	type Solver = SequentialPhragmen<AccountId, sp_runtime::Perbill>;
+	type DataProvider = Staking;
+	type WeightInfo = frame_election_provider_support::weights::SubstrateWeight<Runtime>;
+	type MaxWinners = MaxActiveValidators;
+	type VotersBound = MaxElectingVoters;
+	type TargetsBound = MaxElectableTargets;
+}
+
+pub struct EraPayout<T>(PhantomData<T>);
+impl pallet_staking::EraPayout<Balance> for EraPayout<Staking> {
+	fn era_payout(
+		_total_staked: Balance,
+		_total_issuance: Balance,
+		_era_duration_millis: u64,
+	) -> (Balance, Balance) {
+		let era_index = Staking::active_era().unwrap().index;
+
+		let payout = validators_rewards::era_payout(era_index);
+		let rest = 0;
+
+		(payout, rest)
+	}
+}
